Move spare buffer writer to vortex-buffer
